Move AttributeDimension to its own pair of files.
