+++ src/libs/renderer/renderer/VertexSpecification.cpp
@@ -17,17 +17,6 @@ namespace {
 
 } // anonymous
 
-std::ostream & operator<<(std::ostream & aOut, const AttributeDimension & aAttributeDimension)
-{
-    if (aAttributeDimension.mSecondDimension == 1)
-    {
-        return aOut << aAttributeDimension.mFirstDimension;
-    }
-    else
-    {
-        return aOut << "[" << aAttributeDimension[0] << ", " << aAttributeDimension[1] << "]";
-    }
-}
 
 std::ostream & operator<<(std::ostream &aOut, const AttributeDescription & aDescription)
 {
